refactor(home): align module with frontend code style
