+++ yak-ops-ui/src/pages/home/components/HomeHeader.tsx
@@ -1,0 +1,81 @@
+import { history } from '@umijs/max';
+import { ChevronRight, Database } from 'lucide-react';
+
+import { useHomeHeaderStats } from '../hooks/useHomeHeaderStats';
+
+interface ProfileStatProps {
+  label: string;
+  value: number;
+  arrow?: boolean;
+  onClick?: () => void;
+}
+
+function ProfileStat({ label, value, arrow = false, onClick }: ProfileStatProps) {
+  return (
+    <button
+      type="button"
+      onClick={onClick}
+      className="flex items-center border-0 bg-transparent p-0 text-sm leading-[22px] text-[#747983] transition-colors duration-200 hover:text-[#292c35]"
+    >
+      <span>{label}</span>
+      <strong className="ml-[5px] font-semibold text-[#282b34]">{value}</strong>
+      {arrow ? (
+        <ChevronRight
+          size={14}
+          strokeWidth={1.8}
+          className="ml-[3px] text-[#9599a2]"
+        />
+      ) : null}
+    </button>
+  );
+}
+
+export function HomeHeader() {
+  const stats = useHomeHeaderStats();
+
+  return (
+    <header className="flex h-[116px] items-center px-4">
+      <div className="flex items-center">
+        <div className="flex h-[66px] w-[66px] shrink-0 items-center justify-center overflow-hidden rounded-full border border-white/90 bg-gradient-to-br from-[#dde6ef] via-[#a6c8e2] to-[#5e93d4] text-white/95 shadow-[0_2px_4px_rgba(31,35,41,0.04)]">
+          <Database size={30} strokeWidth={1.5} />
+        </div>
+
+        <div className="ml-4 min-w-0">
+          <div className="flex min-h-[22px] items-center">
+            <span className="whitespace-nowrap text-sm font-medium leading-[22px] text-[#252830]">
+              Yak Ops
+            </span>
+            <span className="mx-3 h-[14px] w-px shrink-0 bg-black/[0.14]" />
+            <span className="whitespace-nowrap text-sm font-normal leading-[22px] text-[#777b84]">
+              Data Operations Platform
+            </span>
+            <span className="mx-3 h-[14px] w-px shrink-0 bg-black/[0.14]" />
+            <span className="whitespace-nowrap text-sm font-normal leading-[22px] text-[#777b84]">
+              让数据接入、开发、编排、质量与消费变得更简单
+            </span>
+          </div>
+
+          <div className="mt-2.5 flex items-center gap-[27px]">
+            <ProfileStat
+              label="数据源"
+              value={stats.dataSourceCount}
+              arrow
+              onClick={() => history.push('/data-source')}
+            />
+            <ProfileStat
+              label="运行中"
+              value={stats.runningCount}
+              arrow
+              onClick={() => history.push('/data-development/executions')}
+            />
+            <ProfileStat
+              label="近7日异常"
+              value={stats.exceptionCount}
+              onClick={() => history.push('/data-development/executions')}
+            />
+          </div>
+        </div>
+      </div>
+    </header>
+  );
+}
