+++ yak-ops-ui/src/pages/home/index.tsx
@@ -1,1 +1,1 @@
-export { default } from './CreatorWorkbenchPage';
+export { default } from './HomePage';
